BCBuilder fix for Dirichlet BCs (#54)
Problem: if a block system contained an empty block and a Dirichlet BC
was imposed that affects the empty block, BCBuilder would just ignore
it. This led to a singular system if the empty block was a diagonal
block.
Fix: BCBuilder fills empty diagonal blocks if a BC has been prescribed
for it.